refactor: move drf views to 'drf' package
